fix bug in indirect indexed y addressing mode
add memory breakpoint support by mouse clicking on the hex dump log window

+++ source/editWindow.cpp
@@ -1695,7 +1695,7 @@ void EditWindow::GotoEmuPC()
 		{
 			if (sf->GetLines().size() > i)
 			{
-				GotoLineCol(i, sf->GetLines()[i]->GetChars().size(), MARK_None, true);
+				GotoLineCol(i, (int)sf->GetLines()[i]->GetChars().size(), MARK_None, true);
 				return;
 			}
 		}
